refactoring: rename misc_helpers -> utils
