+++ sspd/checker/mixed.py
@@ -1,5 +1,5 @@
 from .. import base
-from ..misc_helpers import io
+from ..utils import io
 import os
