+++ sspd/tasks/base.py
@@ -1,6 +1,6 @@
 from .. import base, checker, exceptions
-from ..misc_helpers import io
-from ..misc_helpers.paths import FilePath
+from ..utils import io
+from ..utils.paths import FilePath
 import os
