+++ sspd/utils/ignoring_file.py
